Moved TestableBuilder for ProfileService to UserProfileServiceTest base class. FetchUserProfileInfoTest now extends this base class
